backend/requirements.txt
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
pydantic>=2.0
pydantic-settings>=2.0
python-dotenv>=1.0

langchain>=0.3.0
langchain-core>=0.3.0
langchain-openai>=0.2.0
langchain-chroma>=0.2.0
chromadb>=0.5.0

biopython>=1.80
requests>=2.31.0
tavily-python>=0.5.0
